requirements.txt
argcomplete==2.0.0
aws==0.2.5
bcrypt==3.2.0
boto==2.49.0
boto3==1.21.38
botocore==1.24.38
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
durationpy==0.5
fabric==2.7.0
Flask==2.1.1
Flask-Login==0.6.0
future==0.18.2
hjson==3.0.2
idna==3.3
importlib-metadata==4.11.3
invoke==1.7.0
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
kappa==0.6.0
MarkupSafe==2.1.1
paramiko==2.10.3
pathlib2==2.3.7.post1
placebo==0.9.0
prettytable==3.2.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-slugify==6.1.1
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.0
troposphere==4.0.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.1
wsgi-request-logger==0.4.6
zappa==0.54.1
zipp==3.8.0
